chore: split AMM pages into seperate folders
