+++ src/components/SlippageManager/SlippageManager.style.tsx
