+++ src/components/SlippageManager/SlippageManager.tsx
