+++ src/components/SlippageManager/index.tsx
