+++ src/components/index.tsx
@@ -21,6 +21,8 @@ export { PlusDivider } from './PlusDivider';
 export type { PoolsTableProps } from './PoolsTable';
 export { PoolsTable } from './PoolsTable';
 export { ReceivableAssets } from './ReceivableAssets';
+export type { SlippageManagerProps } from './SlippageManager';
+export { SlippageManager } from './SlippageManager';
 export type { ToastContainerProps } from './ToastContainer';
 export { ToastContainer } from './ToastContainer';
 export * from './TransactionDetails';
