--- src/pages/AMM/Swap/index.tsx
@@ -1,3 +1,0 @@
-import Swap from './Swap';
-
-export default Swap;
