--- src/pages/AMM/shared/components/SlippageManager/index.tsx
@@ -1,2 +1,0 @@
-export type { SlippageManagerProps } from './SlippageManager';
-export { SlippageManager } from './SlippageManager';
