+++ src/pages/Pools/components/PoolsBaseTable/MonetaryCell.tsx
