+++ src/pages/Pools/components/PoolsInsights/PoolsInsights.tsx
@@ -19,7 +19,7 @@ import {
   POOL_CLAIM_REWARDS_FEE_TOKEN_FIELD,
   useForm
 } from '@/lib/form';
-import { calculateAccountLiquidityUSD, calculateTotalLiquidityUSD } from '@/pages/AMM/shared/utils';
+import { calculateAccountLiquidityUSD, calculateTotalLiquidityUSD } from '@/utils/helpers/pool';
 import { AccountPoolsData } from '@/utils/hooks/api/amm/use-get-account-pools';
 import { useGetPrices } from '@/utils/hooks/api/use-get-prices';
 import { Transaction, useTransaction } from '@/utils/hooks/transaction';
