+++ src/pages/Swap/components/SwapLiquidity/SwapLiquidity.tsx
@@ -2,10 +2,9 @@ import { CurrencyExt, LiquidityPool } from '@interlay/interbtc-api';
 
 import { formatUSD } from '@/common/utils/utils';
 import { Card, CardProps, CoinPair, Dd, Dl, DlGroup, Dt, Flex, H2 } from '@/component-library';
+import { calculateTotalLiquidityUSD } from '@/utils/helpers/pool';
 import { DateRangeVolume, useGetDexVolumes } from '@/utils/hooks/api/use-get-dex-volume';
 import { useGetPrices } from '@/utils/hooks/api/use-get-prices';
-
-import { calculateTotalLiquidityUSD } from '../../../shared/utils';
 
 type Props = {
   input: CurrencyExt;
